Two-pass preprocessor, exact implementation of gcc preprocessor.
This commit begins a complete implementation of the gcc two-pass preprocessor, exactly as it is documented. The preprocessor is now a separate pass that lexes the source into tokens, which are then provided to the AST parser in the second pass. Not only does this automatically give us all of the detailed behavior of gcc, parsing source with no preprocessor directives is 30% faster than before!

This commit only implements #define. No ## token pasting or stringification yet.

- Preprocessor directives are checked that they only occur as the first token on a line.
- Self-referential macros are handled correctly, although in extreme cases where the macros refuse to expand because of self-references, the output may not be exactly the same as gcc.
- The parameter count of function macros is checked against the count of args passed in a call.